Move PathCalcParams to its own file
