+++ src/Engine/GameInput/SelectMoveLocation.java
@@ -3,6 +3,7 @@ package Engine.GameInput;
 import java.util.ArrayList;
 
 import Engine.GamePath;
+import Engine.PathCalcParams;
 import Engine.Utils;
 import Engine.XYCoord;
 import Engine.GameInput.GameInputHandler.InputType;
@@ -25,7 +26,7 @@ class SelectMoveLocation extends GameInputState<XYCoord>
   protected OptionSet initOptions()
   {
     // Get valid move locations and return our OptionSet.
-    Utils.PathCalcParams pcp = new Utils.PathCalcParams(myStateData.unitActor, myStateData.gameMap);
+    PathCalcParams pcp = new PathCalcParams(myStateData.unitActor, myStateData.gameMap);
     pcp.start = myStateData.unitCoord;
     pcp.includeOccupiedSpaces = canEndOnOccupied;
     ArrayList<XYCoord> moveLocations = new ArrayList<>(pcp.findAllPaths()); // Need to build a new collection because the typechecker doesn't like children
@@ -113,7 +114,7 @@ class SelectMoveLocation extends GameInputState<XYCoord>
     if( !myStateData.path.getWaypoint(0).GetCoordinates().equals(coord) || !Utils.isPathValid(actor, myStateData.path, myStateData.gameMap, canEndOnOccupied) )
     {
       // The currently-built path is invalid. Try to generate a new one (may still return null).
-      Utils.PathCalcParams pcp = new Utils.PathCalcParams(actor, myStateData.gameMap);
+      PathCalcParams pcp = new PathCalcParams(actor, myStateData.gameMap);
       pcp.start = coord;
       myStateData.path = pcp.findShortestPath(end);
     }
